--- src/classes/location.cpp
@@ -1,16 +1,0 @@
-#include "../../inc/location.hpp"
-
-Location::Location(std::string loc_config) : _config(loc_config)
-{
-
-}
-
-Location::~Location(void)
-{
-
-}
-
-bool	Location::parse_loc(void)	
-{
-	return (true);
-}
